requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
django-environ>=0.11
Pillow>=10.0
drf-spectacular>=0.27
gunicorn>=21.2
whitenoise>=6.6
django-cors-headers>=4.4
celery>=5.4
redis>=5.0